refactor: 메서드 위치 변경, SectionAdditionResult vo->dto 파일 이동
