+++ src/main/java/subway/domain/Sections.java
@@ -7,7 +7,7 @@ import java.util.Objects;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import subway.vo.SectionAdditionResult;
+import subway.dto.SectionAdditionResult;
 
 public class Sections {
