+++ src/main/java/subway/dto/SectionAdditionResult.java
@@ -1,4 +1,4 @@
-package subway.vo;
+package subway.dto;
 
 import java.util.List;
 import java.util.Optional;
