+++ src/test/java/subway/domain/SectionsAddTest.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import subway.vo.SectionAdditionResult;
+import subway.dto.SectionAdditionResult;
 
 class SectionsAddTest {
